Add powerpoint instructions to revision control
Add the full powerpoint deck to revision control for condition
07 instructions.  Added so we have history in revision control
of powerpoint file used to generate the instruction images
from.  Could the generation of the images from the slide deck
be automated?  Also some assets, namely the white shapes, are
not downloading correctly in this current powerpoint deck
for some reason.