Issue #3 stimuli size
Update stimuli size to have 25 degrees for the cue and 12 degrees
in visual field for the stimuli.

Note I am not closing off issue #3 yet.  The cue and stimuli don't
quite exactly match the calculated size from Jonathan on my screen.
Also when I set size in cm/inches explicitly the size doesn't match
on my screen, so I may not 100% understand the issue.  I believe my
monitor size on the development machine may not be registering
correctly?  In any case, it is easy to adjust, lets measure these
physically on the actual experiment monitor and see if they aren't
correct, and if not we can adjust as needed, but they should be
pretty close if not correct for the needed size.