Remove bad test trial
Test subject id 8005 was not completed, and is causing data
workflow to fail.  Removiing this trial.

We may need to make the workflow robust to not fail when intermediate steps
are not building like this, especially for subjects that are not
completed, though once we begin running acutal subjects it shouldn't
be as much of an issue as almost always we will complete the whole
experiment for a subject.