Remove incomplete subject
Not sure if this will be our final convention, but original
subject 0007 had problems and terminated their session
early.  Data files not starting with XXXX 4 digit
subject number will be ignored by the data analysis
workflow scripts.  But we may want to keep data we are
dropping in the repository for reference.  For now
propose prepending data that was incomplete with BAD
so we can keep it but not use it in data analysis.